# apps/core/openapi.yaml
openapi: 3.1.0
info:
  title: Life Engine Core API
  description: >
    JSON REST API for Life Engine Core. Provides data storage (CRUD), full-text
    search, authentication (local-token and OIDC), system health/plugin management,
    quarantine management, Server-Sent Events streaming, and sync conflict resolution.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://127.0.0.1:3750
    description: Default local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/data/{collection}:
    get:
      operationId: listRecords
      summary: List records
      description: Returns a paginated list of records from the named collection.
      tags: [Data]
      parameters:
        - $ref: '#/components/parameters/collection'
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
          description: Number of records to skip.
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 1000 }
          description: Maximum number of records to return.
        - name: sort_by
          in: query
          schema: { type: string }
          description: Field name to sort by (e.g. `created_at`).
        - name: sort_dir
          in: query
          schema: { type: string, enum: [asc, desc], default: asc }
          description: Sort direction.
        - name: filter
          in: query
          schema: { type: string }
          description: URL-encoded JSON filter object. See Data API docs for format.
      responses:
        '200':
          description: Paginated list of records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createRecord
      summary: Create a record
      description: Creates a new record. Core assigns an ID, sets version to 1, and timestamps the record.
      tags: [Data]
      parameters:
        - $ref: '#/components/parameters/collection'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Arbitrary JSON data payload.
      responses:
        '201':
          description: Record created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/data/{collection}/{id}:
    get:
      operationId: getRecord
      summary: Get a record
      description: Returns a single record by its ID.
      tags: [Data]
      parameters:
        - $ref: '#/components/parameters/collection'
        - $ref: '#/components/parameters/recordId'
      responses:
        '200':
          description: The requested record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      operationId: updateRecord
      summary: Update a record
      description: >
        Replaces the data payload of an existing record. Uses optimistic
        concurrency — supply the current version number.
      tags: [Data]
      parameters:
        - $ref: '#/components/parameters/collection'
        - $ref: '#/components/parameters/recordId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecordBody'
      responses:
        '200':
          description: Record updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteRecord
      summary: Delete a record
      description: Permanently deletes a record.
      tags: [Data]
      parameters:
        - $ref: '#/components/parameters/collection'
        - $ref: '#/components/parameters/recordId'
      responses:
        '204':
          description: Record deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---------------------------------------------------------------------------
  # Search API
  # ---------------------------------------------------------------------------
  /api/search:
    get:
      operationId: searchRecords
      summary: Search records
      description: Full-text search across all indexed records with optional collection scoping.
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: Search query string.
        - name: collection
          in: query
          schema: { type: string }
          description: Restrict results to a single collection.
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
          description: Maximum number of results.
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
          description: Number of results to skip.
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Search engine not initialised.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # Auth API — Local Token
  # ---------------------------------------------------------------------------
...
  /api/system/info:
    get:
      operationId: systemInfo
      summary: System info
      description: Returns version, uptime, storage backend, and loaded plugin count.
      tags: [System]
      responses:
        '200':
          description: System information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SystemInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/system/plugins:
    get:
      operationId: listPlugins
      summary: List plugins
      description: Returns a detailed list of all registered plugins and their status.
      tags: [System]
      responses:
        '200':
          description: Plugin list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PluginDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Quarantine endpoints
  # ---------------------------------------------------------------------------
  /api/system/quarantine:
    get:
      operationId: listQuarantine
      summary: List quarantined records
      description: Returns paginated quarantine records.
      tags: [System]
      parameters:
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 1000 }
      responses:
        '200':
          description: Quarantined records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/system/quarantine/{id}/reprocess:
    post:
      operationId: reprocessQuarantine
      summary: Reprocess a quarantined record
      description: >
        Re-validates a quarantined record against its original schema. Restores
        it to the correct collection if validation passes.
      tags: [System]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: The quarantine record UUID.
      responses:
        '200':
          description: Record restored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Record still fails validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/system/quarantine/{id}:
    delete:
      operationId: deleteQuarantine
      summary: Delete a quarantined record
      description: Permanently removes a record from quarantine.
      tags: [System]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: The quarantine record UUID.
      responses:
        '204':
          description: Record deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Events API (SSE)
  # ---------------------------------------------------------------------------
  /api/events/stream:
    get:
      operationId: eventStream
      summary: SSE event stream
      description: >
        Opens a Server-Sent Events connection. Receives real-time events from
        the internal message bus. Keep-alive pings every 15 seconds.
      tags: [Events]
      parameters:
        - name: collection
          in: query
          schema: { type: string }
          description: Only receive `new_records` events for this collection.
        - name: event_type
          in: query
          schema:
            type: string
            enum: [new_records, sync_complete, plugin_loaded, plugin_error]
          description: Only receive events of this type.
      responses:
        '200':
          description: SSE stream opened.
          content:
            text/event-stream:
              schema:
                type: string
                description: >
                  Standard SSE wire format. Event types: new_records,
                  sync_complete, plugin_loaded, plugin_error.
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Conflicts API
  # ---------------------------------------------------------------------------
  /api/conflicts:
    get:
      operationId: listConflicts
      summary: List conflicts
      description: Returns all unresolved sync conflicts with pagination.
      tags: [Conflicts]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 1000 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Conflict list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/conflicts/{id}:
    get:
      operationId: getConflict
      summary: Get a conflict
      description: Returns a single conflict with both versions of the record.
      tags: [Conflicts]
      parameters:
        - $ref: '#/components/parameters/conflictId'
      responses:
        '200':
          description: Conflict details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: dismissConflict
      summary: Dismiss a conflict
      description: Removes a conflict without applying any resolution.
      tags: [Conflicts]
      parameters:
        - $ref: '#/components/parameters/conflictId'
      responses:
        '204':
          description: Conflict dismissed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/conflicts/{id}/resolve:
    post:
      operationId: resolveConflict
      summary: Resolve a conflict
      description: Applies a resolution strategy to a conflict.
      tags: [Conflicts]
      parameters:
        - $ref: '#/components/parameters/conflictId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveConflictBody'
      responses:
        '200':
          description: Conflict resolved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Conflict'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

# =============================================================================
# Components
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Bearer token from `POST /api/auth/token` (local-token mode) or
        `POST /api/auth/login` (OIDC mode).
  parameters:
    collection:
      name: collection
      in: path
      required: true
      schema: { type: string, pattern: "^[a-zA-Z][a-zA-Z0-9_-]*$" }
      description: The collection name (e.g. `tasks`, `contacts`, `emails`).
    recordId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: The record UUID.
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RecordResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Record'
    RecordListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Record'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...